+++ general/alertBox.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
 
 class AlertBox {
-  static showAlertBox(BuildContext context, String title, Widget body,
-      [Function function]) {
+  static showAlertBox(BuildContext context, String title, Widget body) {
     AlertDialog alert = AlertDialog(
       title: Text(title),
       content: body,
@@ -10,7 +9,6 @@ class AlertBox {
         TextButton(
           child: Text("OK"),
           onPressed: () {
-            function();
             Navigator.of(context).pop();
           },
         )
