+++ general/pallete.dart
@@ -1,32 +1,43 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:google_fonts/google_fonts.dart';
 
 const TextStyle kBodyText =
     TextStyle(fontSize: 22, color: Colors.white, height: 1.5);
 
-final appBarTheme = AppBarTheme(centerTitle: false, elevation: 0);
+const Color kWhite = Colors.white;
+const Color kGreen = Colors.green;
 
-const kPrimaryColor = Color(0xFF75c8cc);
-const kSecondaryColor = Color(0xFFaedfdb);
-const kWhite = Color(0xFFf2f2f2);
-const kBlack = Color(0xFF1a1a1a);
-
+const kPrimaryColor = Colors.green;
+const kSecondaryColor = Color(0xFFFE9901);
+const kContentColorLightTheme = Color(0x2F7B7B75);
+const kContentColorDarkTheme = Color(0xFFF5FCF9);
 const kWarninngColor = Color(0xFFF3BB1C);
 const kErrorColor = Colors.redAccent;
 
 const kDefaultPadding = 20.0;
-const kSpacingUnit = 10;
-const kSpacingUnit2 = 14;
 
-final kTitleTextStyle = TextStyle(
-  fontWeight: FontWeight.w600,
-);
+ThemeData mainTheme(BuildContext context) {
+  return ThemeData.dark().copyWith(
+    primaryColor: kPrimaryColor,
+    scaffoldBackgroundColor: kContentColorLightTheme,
+    appBarTheme: appBarTheme,
+    iconTheme: IconThemeData(color: kContentColorDarkTheme),
+    textTheme: GoogleFonts.interTextTheme(Theme.of(context).textTheme)
+        .apply(bodyColor: kContentColorDarkTheme, fontSizeDelta: 2.5),
+    colorScheme: ColorScheme.dark().copyWith(
+      primary: kPrimaryColor,
+      secondary: kSecondaryColor,
+      error: kErrorColor,
+    ),
+    bottomNavigationBarTheme: BottomNavigationBarThemeData(
+      backgroundColor: kContentColorLightTheme,
+      selectedItemColor: Colors.white70,
+      unselectedItemColor: kContentColorDarkTheme.withOpacity(0.32),
+      selectedIconTheme: IconThemeData(color: kPrimaryColor),
+      showUnselectedLabels: true,
+    ),
+  );
+}
 
-final kCaptionTextStyle = TextStyle(
-  fontWeight: FontWeight.w100,
-);
-
-final kButtonTextStyle = TextStyle(
-  fontWeight: FontWeight.bold,
-  color: Colors.white,
-);
+final appBarTheme = AppBarTheme(centerTitle: false, elevation: 0);
