--- general/widgets.dart
@@ -1,5 +1,0 @@
-export 'widgets/alertBox.dart';
-export 'widgets/textInput.dart';
-export 'widgets/buttonRounded.dart';
-export 'widgets/backgroundImage.dart';
-export 'widgets/buttonCircular.dart';
