+++ main.dart
@@ -1,12 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
-import 'package:letbike/account/accountSettings.dart';
 import 'sign/screens.dart';
 import 'app/homePage.dart';
 import 'app/itemPage.dart';
 import 'chat/chatScreen.dart';
-import 'account/accountScreen.dart';
-import 'account/accountSettings.dart';
 
 void main() {
   runApp(MyApp());
@@ -30,9 +27,7 @@ class MyApp extends StatelessWidget {
         "CreateNewAccount": (context) => CreateNewAccount(),
         HomePage.routeName: (context) => HomePage(),
         ItemPage.routeName: (context) => ItemPage(),
-        ChatScreen.routeName: (content) => ChatScreen(),
-        AccountScreen.routeName: (context) => AccountScreen(),
-        AccountSettings.routeName: (context) => AccountSettings()
+        ChatScreen.routeName: (content) => ChatScreen()
       },
     );
   }
