+++ notes.md
@@ -5,25 +5,6 @@
 [Návrh_databáze](https://dbdiagram.io/d/603a99cdfcdcb6230b21cb94)
 
 - [ ] getItem a setItem => itemGet && itemSet
-- [ ] vytvořit general.dart
-    - [ ] export objects.dart
-        - [ ] vztvorit objects.dart
-    - [ ] export widgets.dart
-        - [ ] vztvorit widgets.dart
-            - [x] circularbutton od Lukáše
-            - [ ] imagepicker od Bčka
-            - [ ] roundedButton
-                - [ ] registrace
-                - [ ] login
-            - [x] textinputfield
-                - [x] text
-                - [x] heslo
-                    - [x] zakrýt hesla!
-                - [x] email
-    - [ ] export dbServices.dart
-- [x] snad všechny talčítka: po pushNamed sbalit
-- [x] pošéfovat barvy
-- [ ] reload stránek po popu
 
 ## App Content
 
@@ -80,11 +61,11 @@ http://127.0.0.1/projects/letbike/userRegister.php?username=admin&&email=asd@asd
 
 ### Chat Screen
 
-- [x] skupina __A__
+- [ ] skupina __A__
     - [x] příjem na GitHubu
     - [x] staženo
-    - [x] moje hotovo
-    - [x] master na GitHubu
+    - [ ] moje hotovo
+    - [ ] master na GitHubu
 - [ ] načíst obrázek z databáze
 - [ ] umožnit posílat zprávy menší než 4 znaky
 - [ ] mazat inputfield po odeslání zprávy
@@ -102,7 +83,6 @@ http://127.0.0.1/projects/letbike/userRegister.php?username=admin&&email=asd@asd
 
 - [ ] skupina __A__
     - [x] příjem na GitHubu
-    - [x] staženo
+    - [ ] staženo
     - [ ] moje hotovo
     - [ ] master na GitHubu
-- [x] postavit, Lukáš je kokot
