+++ sign/screens/forgot-password.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import '../../general/pallete.dart';
-import '../../general/widgets.dart';
+import '../widgets.dart';
 
 class ForgotPassword extends StatelessWidget {
   @override
@@ -50,10 +50,9 @@ class ForgotPassword extends StatelessWidget {
                     SizedBox(
                       height: 20,
                     ),
-                    TextInput(
+                    TextInputField(
                       icon: FontAwesomeIcons.envelope,
                       hint: "Email",
-                      identificator: "",
                       inputType: TextInputType.emailAddress,
                       inputAction: TextInputAction.done,
                     ),
