+++ sign/widgets/password-input.dart
@@ -1,0 +1,69 @@
+import 'package:flutter/material.dart';
+import '../../general/pallete.dart';
+
+String _password;
+
+class PasswordInput extends StatelessWidget {
+  const PasswordInput({
+    Key key,
+    @required this.icon,
+    @required this.hint,
+    this.inputType,
+    this.inputAction,
+  }) : super(key: key);
+
+  final IconData icon;
+  final String hint;
+  final TextInputType inputType;
+  final TextInputAction inputAction;
+
+  static String getValue(String hint) {
+    if (hint == "Password") {
+      return _password;
+    } else {
+      return null;
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    Size size = MediaQuery.of(context).size;
+    return Padding(
+      padding: const EdgeInsets.symmetric(vertical: 10.0),
+      child: Container(
+        height: size.height * 0.08,
+        width: size.width * 0.8,
+        decoration: BoxDecoration(
+          color: Colors.grey[500].withOpacity(0.5),
+          borderRadius: BorderRadius.circular(16),
+        ),
+        child: Center(
+          child: TextField(
+            decoration: InputDecoration(
+              border: InputBorder.none,
+              prefixIcon: Padding(
+                padding: const EdgeInsets.symmetric(horizontal: 20.0),
+                child: Icon(
+                  icon,
+                  size: 28,
+                  color: kWhite,
+                ),
+              ),
+              hintText: hint,
+              hintStyle: kBodyText,
+            ),
+            obscureText: true,
+            style: kBodyText,
+            keyboardType: inputType,
+            textInputAction: inputAction,
+            onChanged: (String content) {
+              if (hint == "Password") {
+                _password = content;
+              }
+            },
+          ),
+        ),
+      ),
+    );
+  }
+}
